# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "restaurantAPI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/dishes_structures": {
            "get": {
                "tags": [
                    "DishesStructures"
                ],
                "summary": "Выводит все продукты, входящие в состав блюда",
                "operationId": "dishes_structuresAll",
                "responses": {
                    "200": {
                        "description": "Продукты выведены"
                    },
                    "404": {
                        "description": "Продукты не найдены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "DishesStructures"
                ],
                "summary": "Создает продукт",
                "operationId": "dishes_structuresCreate",
                "responses": {
                    "200": {
                        "description": "Продукт создан"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/dishes_structures/{id}": {
            "get": {
                "tags": [
                    "DishesStructures"
                ],
                "summary": "Выводит продукт по Id",
                "operationId": "dishes_structuresById",
                "responses": {
                    "200": {
                        "description": "Продукт выведен"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "put": {
                "tags": [
                    "DishesStructures"
                ],
                "summary": "Обновляет продукт",
                "operationId": "dishes_structuresUpdate",
                "responses": {
                    "200": {
                        "description": "Продукт обновлен"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "delete": {
                "tags": [
                    "DishesStructures"
                ],
                "summary": "Удаляет продукт по id",
                "operationId": "dishes_structuresDelete",
                "responses": {
                    "200": {
                        "description": "Продукт удален"
                    },
                    "404": {
                        "description": "Продукт не найден"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Вывод всех заказов",
                "operationId": "orderAll",
                "responses": {
                    "200": {
                        "description": "Заказы выведены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Создает заказ",
                "operationId": "orderCreate",
                "responses": {
                    "200": {
                        "description": "Заказ создан"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/orders/{order_id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Вывод заказа по id",
                "operationId": "orderGetById",
                "responses": {
                    "200": {
                        "description": "Заказ выведен"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/orders/close/{order_id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Закрывает заказ",
                "operationId": "orderUpdate",
                "responses": {
                    "200": {
                        "description": "Заказ закрыт"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            }
        },
        "/api/order_dishes": {
            "get": {
                "tags": [
                    "OrderDishes"
                ],
                "summary": "Вывод всех блюд в заказе",
                "operationId": "order_dishesGetAll",
                "responses": {
                    "200": {
                        "description": "Блюда выведены"
                    },
                    "403": {
                        "description": "Нет доступа или пользователь не авторизован"
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderDishes"
                ],
                "summary": "Добавляет блюдо в заказ",
                "operationId": "order_dishesCreate",
                "responses": {
                    "200": {
                        "description": "Блюдо добавлено"
                    },
                    "403": {
                        "description": "Пользователь не авторизован"
                    }
                }
            }
        },
        "/api/dishes/{order_id}/{dishes_id}": {
            "delete": {
                "tags": [
                    "OrderDishes"
                ],
                "summary": "Удаляет блюдо из заказа",
                "operationId": "order_dishesDelete",
                "responses": {
                    "200": {
                        "description": "Блюдо удалено"
                    },
                    "404": {
                        "description": "Блюдо или заказ не найден"
                    },
                    "403": {
                        "description": "Пользователь не авторизован"
                    }
                }
            }
        },
...
    }
}